requirements.txt
nltk==3.8.1
pdfplumber==0.11.4
python-docx==1.1.2
beautifulsoup4==4.12.3
reportlab==4.2.2
streamlit==1.36.0
numpy==2.1.1
pytest==8.3.2
